+++ contract/contract_test.go
@@ -99,7 +99,7 @@ func TestContract_Deploy(t *testing.T) {
 
 	// create an address and fund it
 	key, _ := wallet.GenerateKey()
-	s.Transfer(key.Address(), big.NewInt(1000000000000000000))
+	s.Fund(key.Address())
 
 	p, _ := jsonrpc.NewClient(s.HTTPAddr())
 
@@ -138,7 +138,7 @@ func TestContract_Transaction(t *testing.T) {
 
 	// create an address and fund it
 	key, _ := wallet.GenerateKey()
-	s.Transfer(key.Address(), big.NewInt(1000000000000000000))
+	s.Fund(key.Address())
 
 	cc := &testutil.Contract{}
 	cc.AddEvent(testutil.NewEvent("A").Add("uint256", true))
@@ -163,4 +163,139 @@ func TestContract_Transaction(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Len(t, receipt.Logs, 1)
 	}
+}
+
+func TestContract_CallAtBlock(t *testing.T) {
+	s := testutil.NewTestServer(t, nil)
+	defer s.Close()
+
+	// create an address and fund it
+	key, _ := wallet.GenerateKey()
+	s.Fund(key.Address())
+
+	cc := &testutil.Contract{}
+	cc.AddCallback(func() string {
+		return `
+		uint256 val = 1;
+		function getVal() public view returns (uint256) {
+			return val;
+		}
+		function change() public payable {
+			val = 2;
+		}`
+	})
+
+	artifact, addr := s.DeployContract(cc)
+
+	abi, err := abi.NewABI(artifact.Abi)
+	assert.NoError(t, err)
+
+	contract := NewContract(addr, abi, WithJsonRPCEndpoint(s.HTTPAddr()), WithSender(key))
+
+	checkVal := func(block ethgo.BlockNumber, expected *big.Int) {
+		resp, err := contract.Call("getVal", block)
+		assert.NoError(t, err)
+		assert.Equal(t, resp["0"], expected)
+	}
+
+	// initial value is 1
+	checkVal(ethgo.Latest, big.NewInt(1))
+
+	// send a transaction to update the state
+	var receipt *ethgo.Receipt
+	{
+		txn, err := contract.Txn("change")
+		assert.NoError(t, err)
+
+		err = txn.Do()
+		assert.NoError(t, err)
+
+		receipt, err = txn.Wait()
+		assert.NoError(t, err)
+	}
+
+	// validate the state at different blocks
+	{
+		// value at receipt block is 2
+		checkVal(ethgo.BlockNumber(receipt.BlockNumber), big.NewInt(2))
+
+		// value at previous block is 1
+		checkVal(ethgo.BlockNumber(receipt.BlockNumber-1), big.NewInt(1))
+	}
+}
+
+func TestContract_SendValueContractCall(t *testing.T) {
+	s := testutil.NewTestServer(t, nil)
+	defer s.Close()
+
+	key, _ := wallet.GenerateKey()
+	s.Fund(key.Address())
+
+	cc := &testutil.Contract{}
+	cc.AddCallback(func() string {
+		return `
+		function deposit() public payable {
+		}`
+	})
+
+	artifact, addr := s.DeployContract(cc)
+
+	abi, err := abi.NewABI(artifact.Abi)
+	assert.NoError(t, err)
+
+	contract := NewContract(addr, abi, WithJsonRPCEndpoint(s.HTTPAddr()), WithSender(key))
+
+	balance := big.NewInt(1)
+
+	txn, err := contract.Txn("deposit")
+	txn.WithOpts(&TxnOpts{Value: balance})
+	assert.NoError(t, err)
+
+	err = txn.Do()
+	assert.NoError(t, err)
+
+	_, err = txn.Wait()
+	assert.NoError(t, err)
+
+	client, _ := jsonrpc.NewClient(s.HTTPAddr())
+	found, err := client.Eth().GetBalance(addr, ethgo.Latest)
+	assert.NoError(t, err)
+	assert.Equal(t, found, balance)
+}
+
+func TestContract_EIP1559(t *testing.T) {
+	s := testutil.NewTestServer(t, nil)
+	defer s.Close()
+
+	key, _ := wallet.GenerateKey()
+	s.Fund(key.Address())
+
+	cc := &testutil.Contract{}
+	cc.AddOutputCaller("example")
+
+	artifact, addr := s.DeployContract(cc)
+
+	abi, err := abi.NewABI(artifact.Abi)
+	assert.NoError(t, err)
+
+	client, _ := jsonrpc.NewClient(s.HTTPAddr())
+	contract := NewContract(addr, abi, WithJsonRPC(client.Eth()), WithSender(key), WithEIP1559())
+
+	txn, err := contract.Txn("example")
+	assert.NoError(t, err)
+
+	err = txn.Do()
+	assert.NoError(t, err)
+
+	_, err = txn.Wait()
+	assert.NoError(t, err)
+
+	// get transaction from rpc endpoint
+	txnObj, err := client.Eth().GetTransactionByHash(txn.Hash())
+	assert.NoError(t, err)
+
+	assert.NotZero(t, txnObj.Gas)
+	assert.NotZero(t, txnObj.GasPrice)
+	assert.NotZero(t, txnObj.MaxFeePerGas)
+	assert.NotZero(t, txnObj.MaxPriorityFeePerGas)
 }