+++ jsonrpc/eth.go
@@ -73,16 +73,11 @@ func (e *Eth) GetBlockByHash(hash ethgo.Hash, full bool) (*ethgo.Block, error) {
 
 // GetFilterChanges returns the filter changes for log filters
 func (e *Eth) GetFilterChanges(id string) ([]*ethgo.Log, error) {
-	var raw string
-	err := e.c.Call("eth_getFilterChanges", &raw, id)
-	if err != nil {
+	var logs []*ethgo.Log
+	if err := e.c.Call("eth_getFilterChanges", &logs, id); err != nil {
 		return nil, err
 	}
-	var res []*ethgo.Log
-	if err := json.Unmarshal([]byte(raw), &res); err != nil {
-		return nil, err
-	}
-	return res, nil
+	return logs, nil
 }
 
 // GetTransactionByHash returns a transaction by his hash
@@ -94,16 +89,11 @@ func (e *Eth) GetTransactionByHash(hash ethgo.Hash) (*ethgo.Transaction, error) 
 
 // GetFilterChangesBlock returns the filter changes for block filters
 func (e *Eth) GetFilterChangesBlock(id string) ([]ethgo.Hash, error) {
-	var raw string
-	err := e.c.Call("eth_getFilterChanges", &raw, id)
-	if err != nil {
+	var hashes []ethgo.Hash
+	if err := e.c.Call("eth_getFilterChanges", &hashes, id); err != nil {
 		return nil, err
 	}
-	var res []ethgo.Hash
-	if err := json.Unmarshal([]byte(raw), &res); err != nil {
-		return nil, err
-	}
-	return res, nil
+	return hashes, nil
 }
 
 // NewFilter creates a new log filter
@@ -226,4 +216,54 @@ func (e *Eth) ChainID() (*big.Int, error) {
 		return nil, err
 	}
 	return parseBigInt(out), nil
+}
+
+// FeeHistory is the result of the eth_feeHistory endpoint
+type FeeHistory struct {
+	OldestBlock  *big.Int     `json:"oldestBlock"`
+	Reward       [][]*big.Int `json:"reward,omitempty"`
+	BaseFee      []*big.Int   `json:"baseFeePerGas,omitempty"`
+	GasUsedRatio []float64    `json:"gasUsedRatio"`
+}
+
+func (f *FeeHistory) UnmarshalJSON(data []byte) error {
+	var raw struct {
+		OldestBlock  *ArgBig     `json:"oldestBlock"`
+		Reward       [][]*ArgBig `json:"reward,omitempty"`
+		BaseFee      []*ArgBig   `json:"baseFeePerGas,omitempty"`
+		GasUsedRatio []float64   `json:"gasUsedRatio"`
+	}
+	if err := json.Unmarshal(data, &raw); err != nil {
+		return err
+	}
+	if raw.OldestBlock != nil {
+		f.OldestBlock = raw.OldestBlock.Big()
+	}
+	if raw.Reward != nil {
+		f.Reward = [][]*big.Int{}
+		for _, r := range raw.Reward {
+			elem := []*big.Int{}
+			for _, i := range r {
+				elem = append(elem, i.Big())
+			}
+			f.Reward = append(f.Reward, elem)
+		}
+	}
+	if raw.BaseFee != nil {
+		f.BaseFee = []*big.Int{}
+		for _, i := range raw.BaseFee {
+			f.BaseFee = append(f.BaseFee, i.Big())
+		}
+	}
+	f.GasUsedRatio = raw.GasUsedRatio
+	return nil
+}
+
+// FeeHistory returns base fee per gas and transaction effective priority fee
+func (e *Eth) FeeHistory(from, to ethgo.BlockNumber) (*FeeHistory, error) {
+	var out *FeeHistory
+	if err := e.c.Call("eth_feeHistory", &out, from.String(), to.String(), nil); err != nil {
+		return nil, err
+	}
+	return out, nil
 }